fix(workspace): harden worktree adds and restructure gitworktree adapter
Classify MSYS sh fork crashes (dofork / fork: resource temporarily
unavailable) during git worktree add as transient: force-remove the
partial worktree, prune, and retry exactly once, reusing the branch the
crashed attempt created. Wrap terminal add failures in
ports.ErrWorkspaceProvisionFailed with a bounded git stderr excerpt, and
include the review-translator checkout-hook skip (core.hooksPath).

Also:
- format the branch-conflict path readably on Windows (%s, not %q) so
  ErrBranchCheckedOutElsewhere messages match Windows paths
- consolidate the triplicated add-recovery policy into one
  addWorktreeRecovering helper; existing-branch adds now also recover
  from stale registrations
- move the ErrBranchNotFetched translation into resolveBaseRef
- route temp-index and cherry-pick calls through the shared runners
  (hidden console windows on Windows, stubbable in tests)
- split workspace.go by concern: git.go, paths.go, preserve.go,
  project.go
