+++ backend/internal/adapters/workspace/gitworktree/paths.go
@@ -1,0 +1,174 @@
+package gitworktree
+
+import (
+	"errors"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+
+	"github.com/aoagents/agent-orchestrator/backend/internal/domain"
+	"github.com/aoagents/agent-orchestrator/backend/internal/ports"
+)
+
+func physicalAbs(path string) (string, error) {
+	abs, err := filepath.Abs(path)
+	if err != nil {
+		return "", err
+	}
+	abs = filepath.Clean(abs)
+	if resolved, err := filepath.EvalSymlinks(abs); err == nil {
+		return filepath.Clean(resolved), nil
+	}
+	parent := filepath.Dir(abs)
+	base := filepath.Base(abs)
+	for parent != "." && parent != string(os.PathSeparator) {
+		if resolved, err := filepath.EvalSymlinks(parent); err == nil {
+			return filepath.Join(resolved, base), nil
+		}
+		base = filepath.Join(filepath.Base(parent), base)
+		parent = filepath.Dir(parent)
+	}
+	if resolved, err := filepath.EvalSymlinks(parent); err == nil {
+		return filepath.Join(resolved, base), nil
+	}
+	return abs, nil
+}
+
+// validatePathComponent rejects id values that could escape the managed root
+// once joined into a path. filepath.Join cleans `..` before validateManagedPath
+// runs, so a session id of "../other" would otherwise resolve back inside
+// managedRoot while breaking per-project isolation. Reject any path separator
+// or the special `.`/`..` components at the source.
+func validatePathComponent(name, value string) error {
+	if strings.ContainsAny(value, `/\`) {
+		return fmt.Errorf("%w: %s %q must not contain path separators", ErrUnsafePath, name, value)
+	}
+	if value == "." || value == ".." {
+		return fmt.Errorf("%w: %s %q must not be a path-traversal component", ErrUnsafePath, name, value)
+	}
+	return nil
+}
+
+func (w *Workspace) managedPath(cfg ports.WorkspaceConfig) (string, error) {
+	var path string
+	if cfg.Kind == domain.KindOrchestrator {
+		prefix := resolvedSessionPrefix(cfg)
+		path = filepath.Join(w.managedRoot, string(cfg.ProjectID), "orchestrator", prefix+"-orchestrator")
+	} else {
+		path = filepath.Join(w.managedRoot, string(cfg.ProjectID), string(cfg.SessionID))
+	}
+	return w.validateManagedPath(path)
+}
+
+func (w *Workspace) restorePath(cfg ports.WorkspaceConfig) (string, error) {
+	if cfg.Path != "" {
+		return w.validateManagedPath(cfg.Path)
+	}
+	return w.managedPath(cfg)
+}
+
+// resolvedSessionPrefix returns cfg.SessionPrefix when set, otherwise the first
+// 12 characters of the project ID (matching the display-prefix convention).
+func resolvedSessionPrefix(cfg ports.WorkspaceConfig) string {
+	if p := strings.TrimSpace(cfg.SessionPrefix); p != "" {
+		return p
+	}
+	id := string(cfg.ProjectID)
+	if len(id) <= 12 {
+		return id
+	}
+	return id[:12]
+}
+
+func defaultSessionBranchName(id domain.SessionID) string {
+	return "ao/" + string(id)
+}
+
+func firstNonEmpty(values ...string) string {
+	for _, v := range values {
+		if strings.TrimSpace(v) != "" {
+			return strings.TrimSpace(v)
+		}
+	}
+	return ""
+}
+
+func cleanRelativePath(path string) (string, error) {
+	rel := filepath.ToSlash(strings.TrimSpace(path))
+	if rel == "" {
+		return "", errors.New("relative path is required")
+	}
+	if strings.HasPrefix(rel, "/") {
+		return "", fmt.Errorf("%w: relative path %q must not be absolute", ErrUnsafePath, path)
+	}
+	clean := filepath.ToSlash(filepath.Clean(filepath.FromSlash(rel)))
+	if clean == "." || clean == ".." || strings.HasPrefix(clean, "../") {
+		return "", fmt.Errorf("%w: relative path %q escapes the workspace root", ErrUnsafePath, path)
+	}
+	return clean, nil
+}
+
+func (w *Workspace) validateManagedPath(path string) (string, error) {
+	if path == "" {
+		return "", fmt.Errorf("%w: empty path", ErrUnsafePath)
+	}
+	if !filepath.IsAbs(path) {
+		return "", fmt.Errorf("%w: %q is not absolute", ErrUnsafePath, path)
+	}
+	clean := filepath.Clean(path)
+	if clean != path {
+		return "", fmt.Errorf("%w: %q is not clean", ErrUnsafePath, path)
+	}
+	physical, err := physicalAbs(clean)
+	if err != nil {
+		return "", fmt.Errorf("gitworktree: resolve path %q: %w", path, err)
+	}
+	clean = physical
+	inside, err := pathWithin(w.managedRoot, clean)
+	if err != nil {
+		return "", err
+	}
+	if !inside || clean == w.managedRoot {
+		return "", fmt.Errorf("%w: %q is outside managed root %q", ErrUnsafePath, clean, w.managedRoot)
+	}
+	return clean, nil
+}
+
+func pathWithin(root, path string) (bool, error) {
+	rel, err := filepath.Rel(root, path)
+	if err != nil {
+		return false, fmt.Errorf("gitworktree: compare paths: %w", err)
+	}
+	return rel == "." || (rel != "" && rel != ".." && !strings.HasPrefix(rel, ".."+string(os.PathSeparator))), nil
+}
+
+func pathExistsNonEmpty(path string) (bool, error) {
+	entries, err := os.ReadDir(path)
+	if err == nil {
+		return len(entries) > 0, nil
+	}
+	if errors.Is(err, os.ErrNotExist) {
+		return false, nil
+	}
+	return false, fmt.Errorf("gitworktree: inspect path %q: %w", path, err)
+}
+
+func moveStrayPathAside(path string) (string, error) {
+	for i := 0; i < 100; i++ {
+		candidate := path + ".stray"
+		if i > 0 {
+			candidate = fmt.Sprintf("%s.stray-%d", path, i+1)
+		}
+		if _, err := os.Lstat(candidate); err == nil {
+			continue
+		} else if !errors.Is(err, os.ErrNotExist) {
+			return "", fmt.Errorf("gitworktree: inspect stray destination %q: %w", candidate, err)
+		}
+		if err := os.Rename(path, candidate); err != nil {
+			return "", fmt.Errorf("gitworktree: move stray path %q aside to %q: %w", path, candidate, err)
+		}
+		return candidate, nil
+	}
+	return "", fmt.Errorf("gitworktree: move stray path %q aside: no available destination", path)
+}
